Revert "Merge branch 'server_change_test' of https://github.com/natiiix/KRelay into server_change_test"
This reverts commit 2aeb20176883a8f46676a59ec3e0bf68908512f5.

+++ Lib K Relay/Networking/Packets/Server/ShowEffectPacket.cs
@@ -15,7 +15,7 @@ namespace Lib_K_Relay.Networking.Packets.Server
         public Location PosA;
         public Location PosB;
         public ARGB Color;
-        public float Duration;
+        public double Duration;
 
         public override PacketType Type
         { get { return PacketType.SHOWEFFECT; } }
